Fix missing tests (#3)
* Automated cleanup run

* Formatting, cleanup and removal of some dead code

* Formatting SurveyAnsweringModel.java

* Removal of extraneous lines

* Private field modifiers

* Clean up javadoc and remove some more extra lines

* Basic cleanup

* Add test for QuestionCreationModel

* Add test for SurveyCreationModel

+++ src/main/java/no/hib/megagruppe/webpoll/util/sessionmanager/SeeSurveyOverviewSessionManager.java
@@ -29,7 +29,7 @@ public class SeeSurveyOverviewSessionManager extends SessionManager {
 	public boolean hasSurveyID(){
 		HttpSession session = request.getSession();
 
-		return (Integer) session.getAttribute(SURVEY_ID) != null;
+		return session.getAttribute(SURVEY_ID) != null;
 	}
 	
 	public Timestamp getTimestamp(String daysS, String hoursS, String minutesS){
